renamed more files from index to class name
