+++ 1
@@ -1,0 +1,4 @@
+{
+  "name": "presence",
+  "main": "./presence.js"
+}
