+++ src/utils/flux/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "flux",
+  "main": "./flux.js"
+}
