+++ src/utils/flux/store/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "store",
+  "main": "./store.js"
+}
