+++ src/utils/helpers/chain-functions/__spec__.js
@@ -1,4 +1,4 @@
-import ChainFunction from './index'
+import ChainFunction from './chain-functions'
 
 describe('ChainFunctions', () => {
   let originalFunction, newFunction;
