+++ src/utils/helpers/chain-functions/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "chain-functions",
+  "main": "./chain-functions.js"
+}
