+++ src/utils/helpers/events/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "events",
+  "main": "./events.js"
+}
