+++ src/utils/helpers/immutable/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "immutable",
+  "main": "./immutable.js"
+}
