+++ src/utils/router/__spec__.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import createBrowserHistory from 'history/lib/createBrowserHistory';
-import { startRouter } from './index';
+import { startRouter } from './router';
 
 describe('startRouter', () => {
   let render, router, routes;
