+++ src/utils/router/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "router",
+  "main": "./router.js"
+}
