+++ src/utils/validations/presence/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "presence",
+  "main": "./presence.js"
+}
